Combined Condition1 and Condition2 in new class ConditionTuple: *  encapsulate some of the logic for setting conditions * tuple can be given a name * conditions can be swapped * order of conditions doesn't matter (condition A AND condition B = condition B AND condition A)
This is used for assigning names to condition sets, which are stored in Metadata.

Changed Condition Editor accordingly and added menu entry to menu bar.

At the moment the assigned names are only shown in the mapping list. In the future they will appear as options in the mapping details pane.